orchestrate/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
asn1crypto==1.5.1
backoff==2.2.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==14.0
croniter==2.0.5
cryptography==42.0.7
dagster==1.7.10
dagster-graphql==1.7.10
dagster-pipes==1.7.10
dagster-webserver==1.7.10
docstring_parser==0.16
filelock==3.14.0
fsspec==2024.6.0
gql==3.5.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
grpcio==1.64.1
grpcio-health-checking==1.62.2
h11==0.14.0
httptools==0.6.1
humanfriendly==10.0
idna==3.7
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.1.4
pendulum==3.0.0
platformdirs==4.2.2
protobuf==4.25.3
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.10.1
snowflake-sqlalchemy==1.5.3
sortedcontainers==2.4.0
SQLAlchemy==1.4.52
starlette==0.37.2
structlog==24.2.0
tabulate==0.9.0
time-machine==2.14.1
tomli==2.0.1
tomlkit==0.12.5
toposort==1.10
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
universal_pathlib==0.2.2
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
yarl==1.9.4
